Remove the scheduling_elapsed and scheduling_histogram metrics
They are easily capable of generating literally hundreds of thousands of timeseries,
blocking everything for 10s to thousands of milliseconds.

This leaves the dashboard panels present, mostly because of the time of day, but also
because hopefully we can find a way to trun this back on at least in some limited
capacity.