*: switch to standard crate layout
The Cargo standard for crate layouts looks like this:

    ├── Cargo.toml
    ├── src/
    │   ├── lib.rs
    │   ├── main.rs
    │   └── bin/
    │       ├── named-executable.rs
    │       ├── another-executable.rs
    │       └── multi-file-executable/
    │           ├── main.rs
    │           └── some_module.rs
    ├── benches/
    │   ├── large-input.rs
    │   └── multi-file-bench/
    │       ├── main.rs
    │       └── bench_module.rs
    ├── examples/
    │   ├── simple.rs
    │   └── multi-file-example/
    │       ├── main.rs
    │       └── ex_module.rs
    └── tests/
        ├── some-integration-tests.rs
        └── multi-file-test/
            ├── main.rs
            └── test_module.rs

For historical reasons, our crate layout elides the src directory. This
is a nice simplification when the crate contains only one or two source
files, but scales very poorly as the crate grows benches, tests,
examples, and more modules, as these all get intermingled:

    ├── adts/
    ├── benches/
    ├── examples/
    ├── relation/
    ├── tests/
    ├── Cargo.toml
    ├── a_module.rs
    ├── lib.rs
    └── main.rs

Plus the nonstandard layout is confusing to folks who are used to other
Rust projects.

This patch moves rearranges all of our crates to follow the standard
crate layout.