demo/billing: improve reliability of sink verification
The topic name of the new sink may not be immediately available after
the sink is created; querying for it needs to be wrapped in a retry
loop.

The billing demo previously took a two step approach: first it
queried for the new sink's ID, in a retry loop, and then it queryed for
the topic name using the previously-retried ID, but not in a retry loop.
Instead, just use a SQL join to query for the sink's topic name
directly, and do so in a retry loop.

This fixes some CI flakiness, where the sink demo was incorrectly
failing because it was unable to find the topic name of the new sink.