sql: use datadriven for parser tests
Some of these aren't as clean a win as I was hoping for, since the text
of a parsed AST is pretty noisy to read (I tried pretty-printing them,
but they were so long then that it was a bit obnoxious).

I think this is still a net win in terms of maintainability, but I'm
open to discussing how it could be even better.