coord: absorb sql::Session
The sql crate has now properly abstracted away all details of connection
state that it relies on with the sql::catalog::Catalog trait. Nothing in
the sql crate actually depends on the Session type, which is a clear
sign it belongs in the place it is actually used; i.e., the coordinator.