Push predicate information through views to sources (#3279)
This PR pushes predicate information through views and to sources, which can allow sources to filter information early. This is especially helpful for the upsert operator, which is able to reduce the volume of data that it maintains for correct operation to just those records that pass the filter.

This PR does not actually perform these optimizations for any sources, but it makes the data available.