sql-parser: reorder AST fields to match SQL source order
We try to have the order of fields in the AST match the order they
appear in a SQL statement. This isn't particularly important except
inasmuch as it makes test_basic_visitor easier to follow. Fix a few
places where the AST order did not match the SQL source order.