sql: fix optimization in AggregateFuncVisitor
AggregateFuncVisitor only needs to look at the current level of the
the query, so as an optimization it stops recursing if it sees
Expr::Subquery. There are actually several other forms of subqueries
(Expr::InSubquery, Expr::Exists) that this optimization was not
catching. Teach it to look for Query nodes instead, which is guaranteed
to catch all subqueries.