expr, transform: remove dead binding code
The let/deduplicate transforms have long been disabled due to
performance issues. These transforms are the only dependency on the
`indexmap` crate in our entire codebase, so removing them allows us to
prune that dependency.

We do want let/deduplicate transforms someday, but we'll likely need to
start from scratch to adjust the performance issues. Plus the old
implementations will continue to be available in Git if we need to refer
back to them.