sql: fix comparison semantics for NULL with ANY and ALL
The transformation of IN/ALL/ANY => EXISTS is only valid when both the
LHS and RHS are non-nullable, or when the IN/ALL/ANY clause appears
directly in a WHERE clause. The existing code was only guarded on
whether the right column was non-nullable.

This commit also moves the transformation to a later stage in the
pipeline, after planning but before decorrelation. This makes it
possible to spot when IN/ALL/ANY occurs directly inside of a filter when
nullability does not need to be taken into account. The new
transformation is unfortunately a bit harder to read, largely thanks to Rust pattern
matching struggles.

Note that this change introduced a regression in Q16 of chbench, which
is a NOT IN query where the LHS is nullable but the RHS is non-nullable.
This would seem to require a special null-aware antijoin operator, which
we don't particularly want to introduce at this moment; luckily, we
think the chbench schema intended for the LHS to be non-nullable (it is
in vanilla TPCH), and so this commit just applies that change to the
chbench schema.

Note that this commit presently introduces some regressions, but those
will be resolved once #3449 lands and we can add more aggressive
reduction of the IS NULL operator.

Fix #3319.