coord,sql: don't present unoptimized RelationDesc
We've been bitten several times by the fact that sql::plan::View
presents a RelationDesc for the unoptimized plan. This RelationDesc is
the only source of column names for the view, but duplicates the type
information from the RelationExpr; optimizing the RelationExpr requires
a delicate dance to construct a new RelationDesc with the same column
names but updated types, which several code paths have historically
mishandled.

This commit changes sql::plan::View to contain a Vec<ColumnName> instead
of a RelationDesc. This forces the coord layer to produce a RelationDesc
if it needs one, which will hopefully make it harder to use the types
from the unoptimized plan.