sql: fix nested joins
These got broken in #1426, as apparently we don't have any test coverage
for nested joins. The precise nature of the bug is hard to describe, as
I didn't bother to understand it fully, but in short we were completely
mixing up the join scopes and operators for the nested component.

Roughly, I believe we were planning a query like

    SELECT * FROM a INNER JOIN (b INNER JOIN c ON b.x = c.x) ON a.x = c.x

as

    SELECT * FROM a INNER JOIN b ON a.x = c.x INNER JOIN c ON b.x = c.x

where the most obvious effect of this misplan was an error about c.x not
being in scope.

The correct behavior, implemented by this patch, is as follows. A nested
join stashes away the current stack of joins and builds the nested
component up from scratch, i.e., with nothing in scope and no tablers
join. Only once the nested component is fully built is it joined into
the existing stack of joins.

Fix #3426.