repr: reformulate dummy datums
There are several moments where our optimizer can prove that a column is
not observed, and therefore does not need to be computed. Unfortunately,
our optimizer makes it difficult to actually *remove* this column from
the plan, and so the execution engine inserts a "dummy datum" where it
would have computed the column. If the column is nullable, it uses
`Datum::Null` as the dummy value; otherwise, it uses a reasonable
default value for the column's type, like zero for integers and false
for booleans.

There are two things that concern me about this approach.

The first is that this seems overly concerned about avoiding
`Datum::Null`s in non-nullable columns. If the column truly won't be
observed, then there would be no harm in introducing `Datum::Null` in a
column that is supposed to be non-nullable, because no one will ever
look at that column. If we're worried that the demand optimizations
might be buggy at the moment, well, then shouldn't we be worried about
the fact that we might be blanking out the integers in an observed
column and silently producing wrong data?

The second is that record types, forthcoming in #3505, cannot correctly
implement `ScalarType::dummy_datum` with its current signature, as it is
not possible to construct a dummy datum for a record type with non-zero
arity. E.g., the natural dummy datum for `record(i32, bool)` would be
`Datum::List(0, false)`, but this requires an allocation, which is not
possible given the current signature of ScalarType::dummhy_datum.

So, this commit removes `ScalarType::dummy_datum`. In its place, it
introduces a new `Datum::Dummy` variant, which can be easly constructed
anywhere a placeholder datum is required without knowledge of the
expected type. Crucically, `Datum::Dummy` can never be constructed by a
user, so seeing a `Datum::Dummy` somewhere is a clear sign that an
optimization somewhere is misbehaving, and Materialize can just panic.

I believe the new approach is actually more paranoid in ensuring that
the dummy data is not observed, while also not impeding the
implementation of record types.

Long term, I think we should endeavor to get rid of dummy datums; they
seem to be solely a result of optimizer limitations.