sql: begin introducing more structured errors
More structured errors will allow us to return much more useful errors
to the client, like errors that include a "SQLState" code, a detailed
description, a hint, and so on. Plus, they allow catching and handling
specific errors, which can be useful in planning.

The refactor here is the minimal refactor required to support #1673. The
gist is that we need to catch "unknown column" errors without catching
any _other_ types of errors.

Over time, I expect us to replace all usage of failure::Error in the SQL
planner with this new PlanError type.