sql: fix GROUP BY ordinal references
GROUP BY ordinal references, like the `1` in the following example

    SELECT a, sum(b) FROM foo GROUP BY 1

are meant to refer to an output column, not an input column.

This commit fixes the bug and enables the CockroachDB SLTs that would
have exposed the issue. It also vastly improves the error message when
SELECTing a non-aggregated column.

Fix #443.