transform: fix bug in literal lifting
The LiteralLifting transform was incorrectly hardcoding the output type
of AggregateFunc::{Any,All} as non-nullable, when in fact their output
type is nullable.

This caused incorrect output on the two correlated subquery tests
enabled in this commit.

The fix is to use `AggregateFunc::output_type` to compute the correct
output type; the optimizer should not be duplicating this information.
(There *is* a larger problem of "it is too easy to get nullability
information wrong", but even if that were solved, we still wouldn't want
to be hardcoding information about function output types in the
optimizer.)