doc/user: update docs + release notes for 0.3.1 and 0.4.0
This commit polishes the documentation and release notes for the 0.3.1
and 0.4.0 releases. A bit late, as both of those releases have shipped,
but better late than never.

The big new items of documentation include:

  * basic documentation for record types
  * documentation of escape string literals
  * documentation of typed string literals

I've additionally updated the release notes for 0.3.1 and 0.4.0 to match
the style guide, added many missing release notes, grouped things by
theme where possible, marked backwards incompatible changes, and
resorted several notes that were in the wrong release. These are the
standards I'd like to hold release notes too, and I'm happy to do the
legwork to make that so.