kafka-util: introduce dedicated KafkaAddr type
We were previously using a url::Url to carry around Kafka broker's
host/port information. Unfortunately the Rust URL library refuses to
parse a URL of the form

    12.34.56.78:9092

and instead returns a confusing error about a "relative URL without a
base."

Rather than trying to wrangle this infuriatingly pedantic `url` crate
into dealing with this, introduce a new KafkaAddr type that stores a
host string/port number, and has the flexible parsing we're looking for.