materialized: take responsibility for conditional selection of allocator
The materialized crate should be the one place where we choose which
allocator to use based on the target OS. This commit relieves the prof
crate of that responsibility. One day, hopefully when target-specific
features are supported, the prof crate will be able to expose both
jemalloc- and non-jemalloc-dependent features, but for the moment we
just declare that the prof crate depends on jemalloc.

As an added bonus, with this commit, neither the prof crate nor
the jemallocator crate are linked on macOS. This avoids bringing in the
jemalloc dependency at all on macOS, where it is unused.

This is an alternative to #3854.

Fix #3853.