materialized: change pattern for function handlers
Make HTTP handlers non-async functions, but functions that return an
`impl Future` instead. This is slightly less ergonomic, but it permits
each of these functions to retain a borrow of self, which avoids awkward
and unnecessary cloning of various bits of the Server struct. Instead,
each handler can clone what it needs, if it needs anything at all,
before constructing its 'static future.