catalog: load stored catalog items in gid order
Previously, on restart we loaded catalog items in rowid order. This was fine, as
for most catalog items the rowid order corresponds exactly to the gid order
(between user items). However, some items need to be modified after they have
been created, which we currently do by deleting the old item, and then inserting
a new, modified version of the item. This breaks the rowid ordering.

Note that we will have to revisit this change if we ever persist user items and system
items in the catalog, as the gid order likely does not do what we want between user and
system ids.