repr: improve readability of ColumnType construction
Adjust the ColumnType construction pattern from

    ColumnType::new(ScalarType::Bool, true)

to:

    ScalarType::Bool.nullable(true)

The most important difference is that the purpose of a literal `true` or
`false` is immediately clear to the reader. With the old API, I was
continually forgetting both what the bool parameter meant at all
("nullability"), and then it would take me another moment to remember if
"true" meant "null" or "not null". In the new API, it is much more
obvious.

An alternative to this design would be to introduce an enum like

    enum Nullable {
        Nullable,
        NotNullable
    }

but there are two big downsides with this. The first is the verbosity:
even if you shorten to e.g. `Nullable::Yes`, it's still a mouthful.
The second is that enums don't compose as well as booleans. It's nice to
be able to write

    if ty.nullable { /* ... */ }

or:

    let nullable = ty1.nullable || ty2.nullable;

Using an enum instead of a raw boolean makes such patterns more
difficult.

One downside of the proposed approach is that it's not immediately clear
from the call site that you're creating a `ColumnType`. I think this is
offset by how fundamental the method is; after all, nullability is *the*
difference between a scalar type and a column type.